Update patch set 1
Patch Set 1:

No, the file WCNSS_qcom_cfg.ini is necessary to be present in /system/etc/wifi which it is not in huashan and as well as all devices inheriting blue-common. 
The current path is only /system/etc/firmware/wlan/prima which is required for prima driver but the path /system/etc/wifi is used in wcnss service.

Patch-set: 1