Update patch set 1
Patch Set 1:

Thanks for checking.  You have 3 applications in the SIM card:

      mUiccApplications[0]=APPTYPE_USIM com.android.internal.telephony.uicc.UiccCardApplication@41bbaf78
      mUiccApplications[1]=APPTYPE_ISIM com.android.internal.telephony.uicc.UiccCardApplication@41bfc930
      mUiccApplications[2]=APPTYPE_UNKNOWN com.android.internal.telephony.uicc.UiccCardApplication@41c04a40

The iccid's that are set seem to be associated with mParentApp com.android.internal.telephony.uicc.UiccCardApplication@41bbaf78, the null are associated with @41bfc930.  I am assuming this is normal or harmless, but I don't know a lot about it.  (I only have one, of APPTYPE_SIM, but I think this is because I have an old SIM, which works OK because I don't use mobile data often.)  If your phone number shows correctly in the Settings app, About Phone -> Status, I think it's OK.

Did you test with just #74936 (now merged)?  I guess that would be a good double-check, just to make sure this patch is actually necessary.  I believe #74936 breaks SIM IO for taoshan, then this fixes it.

Patch-set: 1